+++ src/ui.cpp
@@ -58,7 +58,7 @@ void UserInterface::main_loop() {
 
 		if (!network_.time_ok(&last_sync_us)) {
 			set_led(colour::YELLOW);
-		} else if (now_us - transmit_.last_us() > microseconds(1s).count()) {
+		} else if (now_us - transmit_.last_us() > (uint64_t)microseconds(1s).count()) {
 			set_led(colour::RED);
 		} else if (now_us - last_sync_us > 2 * duration_cast<microseconds>(milliseconds(CONFIG_LWIP_SNTP_UPDATE_DELAY)).count()) {
 			set_led(colour::BLUE);
